User should not be able to remove a manager, last manager cannot be unset

+++ backend/src/services/CampaignUserService.js
@@ -109,6 +109,20 @@ class CampaignUserService {
 
         await this.acl.checkUserIsCampaignManager(user.user_id, campaign_id);
 
+        const manager_campaign_user = await this.campaign_user_repository
+            .search({
+                campaign_id,
+                access_level: CampaignUserRepository.MANAGER,
+            });
+
+        if (
+            manager_campaign_user.length === 1
+            && manager_campaign_user.find((item) => item.user_id === id) !== undefined
+            && manager_campaign_user.find((item) => item.user_id === id).access_level !== CampaignUserRepository.MANAGER
+        ) {
+            throw new CustomError(CustomError.UNAUTHORIZED, 'Last manager cannot be unsetted');
+        }
+
         await this.campaign_user_repository
             .update({
                 id,
@@ -187,6 +201,17 @@ class CampaignUserService {
         await this.acl.checkCampaignExists(campaign_id);
 
         await this.acl.checkUserIsCampaignManager(user.user_id, campaign_id);
+
+        const manager_campaign_user = await this.campaign_user_repository
+            .find({
+                campaign_id,
+                id,
+                access_level: CampaignUserRepository.MANAGER,
+            });
+
+        if (manager_campaign_user !== null) {
+            throw new CustomError(CustomError.UNAUTHORIZED, 'Manager user cannot be removed');
+        }
 
         return this.campaign_user_repository.delete(id);
